+++ k8s/src/main/scala/io/buoyant/k8s/package.scala
@@ -1,7 +1,12 @@
 package io.buoyant
 
+import com.fasterxml.jackson.databind.{DeserializationFeature, ObjectMapper}
+import com.fasterxml.jackson.module.scala.DefaultScalaModule
+import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
+import com.twitter.finagle.util.LoadService
 import com.twitter.finagle.{Service, http => fhttp}
 import com.twitter.logging.Logger
+import io.buoyant.config.JsonStreamParser
 
 /**
  * This package contains representations of objects returned by multiple versions of the Kubernetes
@@ -11,4 +16,12 @@ package object k8s {
   type Client = Service[fhttp.Request, fhttp.Response]
 
   private[k8s] val log = Logger.get("k8s")
+
+  val Json = {
+    val mapper = new ObjectMapper with ScalaObjectMapper
+    mapper.registerModule(DefaultScalaModule)
+    mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+    LoadService[SerializationModule].foreach { svc => mapper.registerModule(svc.module) }
+    new JsonStreamParser(mapper)
+  }
 }